python/django-tuts/todo_api/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
